mafatlal/requirements.txt
asgiref==3.8.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
